Test coverage for both legacy and C++20 coroutines
This uncovered the fact that the C++20 tests were not actually
using C++20 coroutines (!) because we were still passing the
`/await` flag, which sends us back to C++17 coroutines.

Updated the `cppwinrt.props` so you can opt into stdcpplatest.
(If you say nothing, then you get the default, which is currently
C++17 with `/await`, but that could change as time passes.)