Avoid spurious WINRT_NATVIS redefinition warning
If you pass `-DWINRT_NATVIS` to the compiler, it defines
the preprocessor symbol to `1`, as if you had written

```cpp
#define WINRT_NATVIS 1
```

This causes debug builds to generate a macro redefinition warning
because the debug build force-defines `WINRT_NATVIS` as the empty
macro.

We don't really care what it defined to, as long as it's defined.
Guard the redefinition to avoid the spurious warning.