Enable weak reference support for pure classic COM objects
Turns out that IWeakReference does work for classic COM:
Even though the output pointer of `IWeakReference::Resolve`
is typed as `IInspectable**`, it is annotated as `iid_is(riid)`,
so COM ignores the formal type and relies on the IID, which
is not constrained to derive from IInspectable. Verified that
the COM marshaling metadata for the Resolve method is unaffected
by the formal type of the output pointer. In particular,
querying the resulting object for `IInspectable` will send a request
over the wire (if never requested before), proving the COM did not
infer `IInspectable` support.

This is a behavior change: Previously, pure classic COM objects
did not support weak references. However, it removes developer
surprise, since they see a method called `get_weak()` and
expect it to work.