[GNUC] Don't use __iso_volatile_load32/64
These intrinsics are missing in mingw-w64 and GCC-compatible compilers.

The __iso_volatile_load intrinsics is intended to tell MSVC to use the
ISO standard C++ non-atomic volatile load, instead of the MSVC extension
which adds other guarantees. For GCC-compatible compilers, the ISO
standard volatile load is already the default, so there is no need to
use these intrinsics.