mingw: Stop using .weak symbols aliases
Weak symbols on mingw-w64 target with GCC and the GNU linker is a bit
broken for the intended use. When using the .weak + .set combo and the
referenced symbol is undefined, LD does not produce an error but instead
treat the referenced symbol like a weak symbol. This can happen if some
of the link library flags is missing. The output executables will crash
on run time, which can be very confusing.

This change makes the WINRT_IMPL_... Windows API declarations on GNUC
use asm labels instead. This means we lose the ability to "hook" these
functions when compiling for mingw-w64 target. But as mentioned in
[#475], developers should not be doing this anyway.

[#475]: https://github.com/microsoft/cppwinrt/pull/475