gcc: Fix `-Wclass-memaccess` warning
`detach_abi(com_array<T>&)` uses two `memset` calls to zero two objects,
one is an `std::pair` (from [#1165]), the other is a `winrt::com_array`
(to clear its data pointer to prevent its destructor from freeing the
array). GCC rightfully warns about this because these types are not
trivially copyable, so calling memset on it is UB.

This change fixes the UB and the GCC warning by reverting the first
`memset` back to using the `std::pair` constructor as before [#1165],
and changing the second `memset` to setting the member fields directly.
Since this requires access to protected members, the function body is
moved into a private member function, then the `winrt::detach_abi` free
function is made `friend` of `com_array` to allow it to call the member
function.

This fix is not applied when `_MSC_VER` is defined in order to preserve
the current behaviour on MSVC, in case [#1165] is still relevant.

[#1165]: https://github.com/microsoft/cppwinrt/pull/1165