Remove unnecessary const qualifier on conversion returns
The conversion from `producer_convert` to `const producer_ref` does not
improve const-correctness, and in compliant compilers¹ prevents the
conversion of `producer_convert` (ala `winrt::implements<D, I>`) to
`producer_ref<T>`'s base type `T`.

In brief², this is because conversion acts as rvalue reference
construction via `T(T&&)`; a `const`-qualified rvalue cannot bind an
rvalue reference.

The `const`ness here doesn't impart any additional safety, either:
- It contains only an untyped pointer, the constness of which does not
  restrict its holder from any additional access to either the
  implementation or projected types.
- It exists to convert to its base type T, and it will do so by copy.
  That copy will contain (hidden) the same pointer, but will often not
  be `const`-qualified; any code that holds it will likely have
  unfettered access to that

In short, it's a `const`-qualified temporary that exists to convert a
`winrt::implements<D, I>` from non-`const` `D` to non-`const` `I`.

Given that holding on to a `producer_ref` is in violation of the
C++/WinRT `impl` namespace contract _anyway_, be it `const`-qualified or
not, this should be a safe change to make.

Closes #1273

¹ http://www.open-std.org/jtc1/sc22/wg21/docs/cwg_active.html#2077
² https://stackoverflow.com/a/70654432