Fix unreliable clock epoch tests
The clock epoch tests are sensitive to CPU speed and
the precise moment the clock ticks over from one second
to another. Change the way we do epoch tests so that
they are not affected by either factor.

This fixes a spurious CI test failure where the test
environment was 3 milliseconds too slow.

https://github.com/oldnewthing/cppwinrt/actions/runs/4195918696/jobs/7276654213