Use safe DLL loading (avoid current directory)
This avoids DLL planting issues during the fallback search
for a base trust server. It uses the `LOAD_LIBRARY_SEARCH_DEFAULT_DIRS`
flag, which searches the usual places, including the application directory,
but exclude the current directory. The flag is supported on Win8 onward,
as well as on Win7 if you have KB2533623 installed (which you really should).
If you need to support versions of Win7 that haven't been patched
since 2011, then stick with whatever version of C++/WinRT you are
currently using.