resume_agile to allow coroutine to resume in any apartment
By default, `co_await`'ing a Windows Runtime asynchronous operation
resumes in the same COM apartment. We provide a way to override
this behavior and resume in any apartment by wrapping the
asynchronous operation in `resume_agile()`.

```cpp
extern IAsyncAction Example();

// This resumes in the same apartment
co_await Example();

// This resumes in any apartment
co_await resume_agile(Example());
```

The guts of the change are in `disconnect_aware_handler`
which accepts a `preserve_context` template parameter,
defaulting to `true`. If `false`, then we bypass the
`resume_apartment()` on completion. The `false` value
is provided by the `resume_agile()` function.

To remove the `resume_apartment_context` from the
`disconnect_aware_handler` in the case where we
don't need it, I use EBO and make it a base class,
so that it disappears when an empty `ignore_apartment_context`
is used.

While I was there, I introduced `movable_primitive<T>` which
provides RAII-like functionality to scalars like integers
and pointers. This allows us to simplify `disconnect_aware_handler`
and `resume_apartment_context`, which previously had to override
all the copy and move operations in order to reset the scalar
on move.

Also fix the `await_adapter.cpp` tests so they don't leak DispatcherQueues,
by explicitly shutting down all the dispatcher queues we created.
This required moving all the controllers to top-level so we can shut
them down after the tests have completed.