Add get_iids_tearoff() hook.
Add a `get_iids_tearoff()` hook to complement the `query_interface_tearoff()`
hook.

`query_interface_tearoff()` allows dynamically extending types with
additional interfaces that can be accessed with the `QueryInterface()`
method. However, these interfaces were not discoverable by the `GetIids()`
method (and therefore not by the `winrt::get_interfaces()` function
either).

The `get_iids_tearoff()` hook allows adding the GUIDs of interfaces
to the array returned by `GetIids()`.

This also fixes a bug in the implementation of the
`root_implements_type::is_composing` branch of `NonDelegatingGetIids()`
where `*array` was updated causing the local iids to be unreachable by
the caller and risking the caller reading past the end of the array.