All cppwinrt vcxproj should build with /W4 and /WX (warnings turned up, treated as error).  Fix existing violations. (#1487)
Why is this change being made?
The most recent PR introduced a warning for non-Clang consumers. Many projects build with warnings-as-errors so this is a build break for those projects. Because cppwinrt is widely used by many such projects it should itself build with warnings as errors enabled so that this doesn't happen again in the future. Plus it is general goodness to build with warnings cranked up and blocking.

Briefly summarize what changed
I added /WX to every single vcxproj in this repo (which required a whole lot of redundant edits). Many of them didn't have /W4 so I added it wherever it was missing. The project files definitely need some refactoring to reduce duplication. But for this change I decided to modify them as-is without also including a refactoring.

I also fixed the newly-blocking warnings. MSVC only had the one new warning to fix.  Clang had a whole bunch of previously-unknown warnings that needed to be fixed or suppressed to get the build passing again.

I also removed a typo'd (warning is singular, not plural) that was trying to disable warnings in one specific project. It has no warnings so we might as well baseline them as blocking.

I also added a new clang-only test file to try and exercise the LTO public visibility annotation.

How was this change tested?
It builds locally (x64 Debug+Release, MSVC + Clang).